unity-app/tests/integration/requirements.txt
python-can>=4.0.0
python-can-isotp>=2.0.0